Main UI Refactoring.
