+++ src/screen/Main/exData.js
@@ -1,0 +1,26 @@
+var exampleData = [
+    {
+        name: "자전거타자전거타자전거타자전거타",
+        dairyID: "231",
+        date: "2020.12.29",
+        writer: "nonokee",
+        imgSrc: "../../../resource/Temp/orange.png",
+        hashTag: ["정보", "19", "한강"],
+        point: 10000,
+        like: 230,
+        dislike: 12
+    },
+    {
+        name: "빵터지는 최불암 깔깔 유모아",
+        dairyID: "232",
+        date: "2020.12.29",
+        writer: "콩수내비",
+        imgSrc: null,
+        hashTag: ["유머", "최불암"],
+        point: 1900,
+        like: 130,
+        dislike: 172
+    }
+]
+
+export {exampleData};
